frontend/references/september-main/pyproject.toml
[project]
name = "voice-compliance-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "elevenlabs>=2.30.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
